Really fix MSVC warning...
git-svn-id: https://svn.osgeo.org/gdal/trunk@37695 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/frmts/gtiff/geotiff.cpp
@@ -7470,7 +7470,7 @@ void GTiffDataset::FillEmptyTiles()
                                                     TIFFClientdata( hTIFF ));
                         const vsi_l_offset nCurOffset = VSIFTellL(fp);
                         VSIFSeekL(fp, nOffset, SEEK_SET);
-                        VSIFReadL(pabyRaw, 1, nRawSize, fp);
+                        VSIFReadL(pabyRaw, 1, static_cast<int>(nRawSize), fp);
                         VSIFSeekL(fp, nCurOffset, SEEK_SET);
                     }
                 }
